June 27, 2026 @ 21:25
NOTE TO SELF:
Deferred messages are not being deleted properly. See: DiscordMessage#noMessage(IReplyCallback, boolean)

+++ FireGenDiscord/src/main/java/net/noahf/firegen/discord/utilities/DiscordMessages.java
@@ -74,8 +74,9 @@ public class DiscordMessages {
                 .queueAfter(selfDestructAfter, TimeUnit.SECONDS);
     }
 
-    public static void noMessage(IReplyCallback event) {
+    public static void noMessage(IReplyCallback event, boolean deleteIfExists) {
         if (event.isAcknowledged()) {
+            if (!deleteIfExists) return;
             event.getHook().deleteOriginal().queue();
         } else {
             event.reply(" ").setEphemeral(true).complete().deleteOriginal().queue();
